+++ src/components/Map/index.ts
@@ -1,2 +1,4 @@
 export { Map } from './Map'
 export { CountryLayer } from './CountryLayer'
+export { DragHandler } from './DragHandler'
+export { InteractiveOverlay } from './InteractiveOverlay'